Add view() method in Table and fix bug
I added the view() method, which views one or more tables (Name of table, names of columns, data).
Made an extra method that finds the maximum length of a row in each column to set the width of each column based on its maximum field width.
Fixed a bug where the name of the table was stored as the name of all the variables.
Fixed a bug where the linesNumber in every row was the same (=1).
Adjusted the main() method in respect to the alterations to test them.