Add full view function
View table(s). View attribute(s) in one table. Also added methods that find the position of one table in the tables field, multiple tables in the tables field, attribute(s) in one table. In the signature of the method that views attributes, there's a small hack that allowed me to continue without changing my code completely > added a placebo boolean parameter that is useless, just to have a different signature. It's completely useless, true or false, same shit. Just don't forget to add that if ever you write code containing that method.