Alter viewTable in a manner that it accepts one table only
I thought about it and decided to change the design a little bit. Apparently, it will be more practical to utilize viewing multiple tables on the app level, rather than the called method. So from now on we will call viewTable on a single table name. If multiple tables are given by the user to be shown, then just utilize a loop on the app level.