Major bug fixes
There is a Debugging.txt file in the parent folder with a detailed report of bugs found during manual tests, meaning the tests you have developed are highly deficient. BUgs that are already fixed are marked with 3 stars before and after the corresponding line in the file. Still working on fixing all of those. You'll have to work on re-building your tests based on the bug report. Thanks