Fix checkAnswer bug and delete table with save database combo bug
In checkAnswer method using || was producing wrong results. Deleting a table and saving the database would not produce the expected results if the database had been previously saved with the deleted table. All resources processing the files had to be closed, in order for them to be deleted, and the ones that corresponded to existing files to be rewritten