Add cancel EVERY-DAMN-WHERE, and more.
I added the choice for everyone to type cancel whenever they want to return to the above level menu if they mistakenly entered a mode they did not want. I still need to do some refinements and additions, but it is functional for now. Also adjusted the delete mode menus. We had a method that was not used, fixed some outputs etc. There was a bug where we deleted a database and the databaseNumber was not reduced. Also fixed. Adjusted the setDatabaseNumber method using a manually inserted margin instead of just adding one (we needed to put -1 as an interval when a database was deleted). Finally, I deleted App.java, AppTest.java and travis.yml, all unused files.